I found and solved a bug. The bug is not visible while running the project, but when the debugger from qtcreator is used (at least with tcl84), closing the window triggers a segfault.
I solved this bug (which is the same the case described here: http://coding.derkeiler.com/Archive/Tcl/comp.lang.tcl/2006-10/msg01192.html )
by implementing the Tcl_DriverCloseProc for the two Tcl_ChannelType define in qtclconsole.cpp
quote from the tcl documentation providing details for TCL_CHANNELTYPE  http://tclhelp.net/TclLib/CrtChannel.htm
"It is not necessary to provide implementations for all channel operations. Those which are not necessary may be set to NULL in the struct: blockModeProc, seekProc, setOptionProc, getOptionProc, and close2Proc, in addition to flushProc, handlerProc, threadActionProc, and truncateProc. Other functions that cannot be implemented in a meaningful way should return EINVAL when called, to indicate that the operations they represent are not available."

Not wanting to venture into the typedef, I declared the function as a global function. This should be placed in a more meaningful place if necessary. I don't know where the best place would be.

This is my last commit (from tbelair@miranda.com) All questions and comments should be redirected to smartine@miranda.com
It was a pleasure contributing to this worthwhile project.